backend/requirements.txt
flask
opencv-python-headless
ultralytics
supervision
numpy==1.25.2
pymongo
python-dotenv
